1143.Longest-Common-Sub-Sequence/Solution_test.go
package leetcode

import (
	"testing"
)

func Test_longestCommonSubsequence(t *testing.T) {
	type args struct {
		text1 string
		text2 string
	}
	tests := []struct {
		name string
		args args
		want int
	}{
		{
			"test-1",
			args{"abcde", "ace"},
			3,
		},
		{
			"test-2",
			args{"abc", "abc"},
			3,
		},
		{
			"test-3",
			args{"abc", "def"},
			0,
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			if got := longestCommonSubsequence(tt.args.text1, tt.args.text2); got != tt.want {
				t.Errorf("longestCommonSubsequence() = %v, want %v", got, tt.want)
			}
		})
	}
}
